refactor: rename endpoints.js to host-rules.js
