+++ lib/config/index.js
@@ -12,7 +12,7 @@ const { getPlatformApi } = require('../platform');
 const { resolveConfigPresets } = require('./presets');
 const { get, getLanguageList, getManagerList } = require('../manager');
 
-const endpoints = require('../util/endpoints');
+const endpoints = require('../util/host-rules');
 
 exports.parseConfigs = parseConfigs;
 exports.mergeChildConfig = mergeChildConfig;
