+++ lib/datasource/docker.js
@@ -4,7 +4,7 @@ const URL = require('url');
 const parseLinkHeader = require('parse-link-header');
 const wwwAuthenticate = require('www-authenticate');
 const { isVersion, sortVersions } = require('../versioning/docker');
-const endpoints = require('../util/endpoints');
+const endpoints = require('../util/host-rules');
 
 module.exports = {
   getDigest,
