+++ lib/datasource/npm.js
@@ -9,7 +9,7 @@ const registryAuthToken = require('registry-auth-token');
 const parse = require('github-url-from-git');
 const { isBase64 } = require('validator');
 const { isVersion, sortVersions } = require('../versioning')('semver');
-const endpoints = require('../util/endpoints');
+const endpoints = require('../util/host-rules');
 
 module.exports = {
   maskToken,
