+++ lib/manager/composer/lock-file.js
@@ -2,7 +2,7 @@ const { exec } = require('child-process-promise');
 const fs = require('fs-extra');
 const upath = require('upath');
 
-const endpoints = require('../../util/endpoints');
+const endpoints = require('../../util/host-rules');
 
 module.exports = {
   getLockFile,
