+++ lib/platform/bitbucket/bb-got-wrapper.js
@@ -1,6 +1,6 @@
 const got = require('got');
 const URL = require('url');
-const endpoints = require('../../util/endpoints');
+const endpoints = require('../../util/host-rules');
 
 let cache = {};
