+++ lib/platform/bitbucket/index.js
@@ -1,7 +1,7 @@
 const parseDiff = require('parse-diff');
 const api = require('./bb-got-wrapper');
 const utils = require('./utils');
-const endpoints = require('../../util/endpoints');
+const endpoints = require('../../util/host-rules');
 
 let config = {};
