+++ lib/platform/github/gh-got-wrapper.js
@@ -2,7 +2,7 @@ const URL = require('url');
 const ghGot = require('gh-got');
 const delay = require('delay');
 const parseLinkHeader = require('parse-link-header');
-const endpoints = require('../../util/endpoints');
+const endpoints = require('../../util/host-rules');
 
 let cache = {};
 let stats = {};
