+++ lib/platform/github/index.js
@@ -5,7 +5,7 @@ const showdown = require('showdown');
 const URL = require('url');
 
 const get = require('./gh-got-wrapper');
-const endpoints = require('../../util/endpoints');
+const endpoints = require('../../util/host-rules');
 const Storage = require('./storage');
 const GitStorage = require('../git/storage');
