+++ lib/platform/gitlab/gl-got-wrapper.js
@@ -3,7 +3,7 @@ const glGot = require('gl-got');
 const delay = require('delay');
 
 const parseLinkHeader = require('parse-link-header');
-const endpoints = require('../../util/endpoints');
+const endpoints = require('../../util/host-rules');
 
 let cache = {};
