+++ lib/platform/gitlab/index.js
@@ -3,7 +3,7 @@ const addrs = require('email-addresses');
 const showdown = require('showdown');
 
 const get = require('./gl-got-wrapper');
-const endpoints = require('../../util/endpoints');
+const endpoints = require('../../util/host-rules');
 
 const converter = new showdown.Converter();
 converter.setFlavor('github');
