+++ lib/platform/vsts/index.js
@@ -1,7 +1,7 @@
 // @ts-nocheck //because of logger, we can't ts-check
 const vstsHelper = require('./vsts-helper');
 const vstsApi = require('./vsts-got-wrapper');
-const endpoints = require('../../util/endpoints');
+const endpoints = require('../../util/host-rules');
 
 const config = {};
