+++ lib/platform/vsts/vsts-got-wrapper.js
@@ -1,5 +1,5 @@
 const vsts = require('vso-node-api');
-const endpoints = require('../../util/endpoints');
+const endpoints = require('../../util/host-rules');
 
 module.exports = {
   vstsObj,
