+++ lib/workers/pr/changelog/source-github.js
@@ -1,5 +1,5 @@
 const URL = require('url');
-const endpoints = require('../../../util/endpoints');
+const endpoints = require('../../../util/host-rules');
 const versioning = require('../../../versioning');
 const ghGot = require('../../../platform/github/gh-got-wrapper');
 const { addReleaseNotes } = require('./release-notes');
