+++ test/datasource/docker.spec.js
@@ -1,9 +1,9 @@
 const got = require('got');
 const docker = require('../../lib/datasource/docker');
-const endpoints = require('../../lib/util/endpoints');
+const endpoints = require('../../lib/util/host-rules');
 
 jest.mock('got');
-jest.mock('../../lib/util/endpoints');
+jest.mock('../../lib/util/host-rules');
 
 describe('api/docker', () => {
   describe('getDigest', () => {
